[NaCl] Fix memory corruption in truetype rendering.
Check for the case where a glyph attempts to draw out side
the allocated region.  In theory this check should not be
needed but these seems to be certain font sizes that result
in certain glyphs having a larger horiBearingY than the
font's ascender.

Review URL: https://codereview.chromium.org/12392002