Add touch ID to lua touch handler
This means that lua touch handlers have access to the touchid
and can return false if they don't wish to handle a certain
touch.  This is useful for multi-touch, and also for hybrid
apps where some touches are handles by lua and some by C++.

This does not effect the tolua bindings or the existing lua
code as the extra argument can be silently ignored.