Add CCPhysicsNode
This allows nodes other than sprites to be be linked to
physics bodies.  In particular it allows a compound
object hierarchy the might contain many sprites or other
types of node, to be linked to single physics body by
the normal cocos parent relationship without mandating that
the root of the hierarchy be a sprite.

R=binji@chromium.org

Review URL: https://codereview.chromium.org/15067017